requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-file==1.0.1
django-file-md5==1.0.3
django-simple-history==3.0.0
djangorestframework==3.13.1
drf-yasg==1.20.0
hurry.filesize==0.9
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
packaging==21.3
pdf2image==1.16.0
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.0
psutil==5.9.0
pycparser==2.21
pycryptodome==3.14.1
PyMuPDF==1.19.5
pyparsing==3.0.7
PyPDF2==1.26.0
pytesseract==0.3.8
python-decouple==3.5
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
Wand==0.6.7
